# bundles/marketdataprovider/de.hf.dac.marketdataprovider.restservice/src/main/java/de/hf/dac/marketdataprovider/restservice/marketdataresources/MDRunnerResource.java
/** ----------------------------------------------------------------------------
 *
 * ---          HF - Application Development                       ---
 *              Copyright (c) 2014, ... All Rights Reserved
 *
 *
 *  Project     : dac
 *
 *  File        : MDRunnerResource.java
 *
 *  Author(s)   : hf
 *
 *  Created     : 17.02.2017
 *
 * ----------------------------------------------------------------------------
 */

package de.hf.dac.marketdataprovider.restservice.marketdataresources;


import com.google.gson.Gson;
import de.hf.dac.api.io.routes.job.JobInformation;
import de.hf.dac.api.io.routes.job.WrappedJobParameter;
import de.hf.dac.api.security.SecuredResource;
import de.hf.dac.marketdataprovider.api.application.MarketDataEnvironment;
import de.hf.dac.marketdataprovider.api.application.OpLevel;
import de.hf.dac.marketdataprovider.api.application.OpType;
import de.hf.dac.marketdataprovider.api.runner.BaseMDRunnerParameter;
import de.hf.dac.marketdataprovider.api.runner.ImportRunnerParameter;
import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
import org.apache.http.HttpStatus;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import javax.ws.rs.Consumes;
import javax.ws.rs.POST;
import javax.ws.rs.Path;
import javax.ws.rs.PathParam;
import javax.ws.rs.Produces;
import javax.ws.rs.core.MediaType;
import javax.ws.rs.core.Response;

@Api(tags = "Jobs")
public class MDRunnerResource extends SecuredResource<OpType,OpLevel> {
    private static final Logger LOG = LoggerFactory.getLogger("MDRunnerResource");
    final protected static Gson gson = new Gson();
    final protected MarketDataEnvironment marketDataEnvironment;


    public MDRunnerResource(MarketDataEnvironment marketDataEnvironment) {
        super(marketDataEnvironment);
        this.marketDataEnvironment = marketDataEnvironment;
    }

    @POST
    @Path("/{jobtype}/start")
    @ApiOperation(response = JobInformation.class, value = "execute marketdata launcher", notes = "Execute marketdata Core Launcher", nickname = "VERBATIMstart")
    @ApiResponses(value = { @ApiResponse(code = HttpStatus.SC_OK, message = "Job Submitted", response = JobInformation.class) })
    @Produces(MediaType.APPLICATION_JSON)
    @Consumes(MediaType.APPLICATION_JSON)
    public Response start(@PathParam("jobtype") @ApiParam(value="The JobType") String jobType,
            @ApiParam(name = "params", value = "Parameter") BaseMDRunnerParameter params) {
        //checkOperationAllowed(OpType.WRITE, jobtype);
        //todo wird hier env benötigt?, sollte nicht besser der Jobtype mitgegeben werden
        JobInformation jobInformation = marketDataEnvironment.getDispatcher()
            .sendJob(new WrappedJobParameter(params, "dev", null, WrappedJobParameter.RUNNER_REQUEST, WrappedJobParameter.RUNNER_RESULT));

        return Response.ok(gson.toJson(jobInformation)).build();
    }
}