# tools/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Dac Service REST API",
    "version" : "1.1",
    "title" : "Dac Services",
    "contact" : {
      "name" : "HF"
    },
    "license" : {
      "name" : "Apache 2.0 License",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath" : "/dac/rest",
...
  "paths" : {
...
    "/runner/{env}/{jobtype}/start" : {
      "post" : {
        "tags" : [ "MDRunner", "Jobs" ],
        "summary" : "execute marketdata launcher",
        "description" : "Execute marketdata Core Launcher",
        "operationId" : "VERBATIMstart",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "jobtype",
          "in" : "path",
          "description" : "The JobType",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "params",
          "description" : "Parameter",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/BaseMDRunnerParameter"
          }
        }, {
          "name" : "env",
          "in" : "path",
          "description" : "The Service Environment",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Job Submitted",
            "schema" : {
              "$ref" : "#/definitions/JobInformation"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "JobInformation" : {
      "type" : "object",
      "properties" : {
        "uuid" : {
          "type" : "string"
        },
        "startTime" : {
          "type" : "string"
        },
        "endTime" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string",
          "enum" : [ "UNKNOWN", "QUEUED", "PROCESSING", "FINISHED", "FAILED" ]
        },
        "statusChangeHistory" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/StatusChange"
          }
        },
        "payload" : {
          "type" : "object"
        },
        "result" : {
          "type" : "object"
        },
        "statusMessageCollector" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/StatusChange"
          }
        }
      }
    },
    "StatusChange" : {
      "type" : "object",
      "properties" : {
        "changeTime" : {
          "type" : "string"
        },
        "oldStatus" : {
          "type" : "string",
          "enum" : [ "UNKNOWN", "QUEUED", "PROCESSING", "FINISHED", "FAILED" ]
        },
        "newStatus" : {
          "type" : "string",
          "enum" : [ "UNKNOWN", "QUEUED", "PROCESSING", "FINISHED", "FAILED" ]
        },
        "changeInformation" : {
          "type" : "string"
        }
      }
    },
...
    "BaseMDRunnerParameter" : {
      "type" : "object",
      "properties" : {
        "params" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        },
        "environment" : {
          "type" : "string"
        }
      }
    },
...
  }
}